Edit forest plot

This function is used to edit the graphical parameters of text and background of the forest plot.

Usage

edit_plot(
  plot,
  row = NULL,
  col = NULL,
  part = c("body", "header"),
  which = c("text", "background", "ci"),
  gp = gpar(),
  ...
)

Arguments

plot

A forest plot object.

row

A numeric value or vector indicating row number to edit in the dataset. Will edit the whole row if left blank for the body. This will be ignored if the part is "header".

col

A numeric value or vector indicating column to edit in the dataset. Will edit the whole column if left blank.

part

Part to edit, "body" (default) or "header".

which

Which element to edit, "text", "background" or "ci" (confidence interval). This will not edit diamond shaped summary CI, please change it with forest_theme. Also, change in ci will not have any impact on the legend.

gp

Pass gpar parameters, see gpar. It should be passed as gpar(col = "red"). For which = "ci", please refer to forest_theme ci_* parameters for the editable elements.

...

Other parameters to be passed to the grobs. See textGrob for the "text" part and rectGrob for "background". This is ignored when which = "ci" because non-graphical parameters cannot be changed for the confidence interval.

Value

A gtable object.

See also